requirements.txt
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
pandas>=1.5.0
requests>=2.28.0
python-dotenv>=0.20.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
streamlit>=1.28.0
plotly>=5.17.0 